airflow/requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
boto3
pandas
dbt-core
dbt-snowflake
python-dotenv